FILE: moves reduce-recursively into internal folder
