+++ libs/core/util-selection/src/lib/internal/reduce-recursively.ts
